feat(home): refactor home page into multitple view
